move New* functions beneath the types they construct
